requirements.txt
# Core AI
anthropic>=0.40.0

# PDF processing
pymupdf>=1.24.0
pdfplumber>=0.11.0

# Office formats (Phase 2)
python-docx>=1.1.0
python-pptx>=1.0.0
openpyxl>=3.1.0

# File type detection
python-magic>=0.4.27

# Vector database
chromadb>=0.5.0

# Image handling
Pillow>=10.0.0

# Structured data
pyyaml>=6.0

# Utilities
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
pytest-cov>=5.0.0
